Initial support for database clustering in Horizon
Added a separate panel for clusters.

This panel contains the Clusters Table with table actions to
Launch and Terminate clusters.  There are row actions Add Shard
and Reset Password (and their associated dialogs) that are specific
to MongoDB and Vertica respectively.

The Clusters Details will include the following tabs:
- Overview
- Instances (table of instances belonging to this cluster)

The launch panel has custom fields for MongoDB and Vertica.
The custom fields will be dynamically shown based on the datastore
selected.

Added a db_capability utility to aid in identifying the specific
datastores.

Added network selection dropdown if neutron is enabled.

Change-Id: I3711766773952d53f20788d08aeedf8fd8139547
Co-Authored-By: Saurabh Surana <saurabh.surana@hp.com>
Co-Authored-By: Duk Loi <duk@tesora.com>
Co-Authored-By: Anna Shen <ruiyuan.shen@hp.com>
Implements: blueprint database-clustering-support