Initial support for database clustering in Horizon
Added a separate panel for clusters.

This panel contains the Clusters Table with table actions to
Launch and Terminate clusters.  There are row actions Add Shard
and Reset Password (and their associated dialogs) that are specific
to MongoDB and Vertica respectively.

The Clusters Details will include the following tabs:
- Overview
- Instances (table of instances belonging to this cluster)

The launch panel has custom fields for MongoDB and Vertica.
The custom fields will be dynamically shown based on the datastore
selected.

Added a db_capability utility to aid in identifying the specific
datastores.

Co-Authored-By: Duk Loi <duk@tesora.com>
Co-Authored-By: Anna Shen <ruiyuan.shen@hp.com>
Change-Id: I047f4d37449070adfd0ea66ad010982f35c049aa
Implements: blueprint database-clustering-support