Change in-memory config group to using a cache
The original implementation of config groups uses an in memory version
of the config.  This is only good within a single process in the web
server.  The apache server can use multiple processes so it was
necessary to use a cache that can be accessible across processes.

Made changes to the config_param_manager to support using a cache.

Change-Id: I3444cc5a0ff6f6cc712887e1baea6386d0bf0c85
Closes-Bug: DBAAS-296