Support operations on a trove instance user
Added a table action on the users table to display the
create user dialog.

Added a row action on the users table to display the
edit user information dialog.

Added a row action on the users table to display the
access details panel as a list of databases on the instance
along with a column indicating whether the user has access
to the database.  There are also row actions that are enabled
for granting or revoking access to the database.

Added unit tests for the new dialogs and panel.

Modified the api to support the new user actions.

Change-Id: Ice2425a614a6e643f55149c46190d5f3ef46c299
Implements: blueprint trove-instance-user-support