Updated theme for the accordion bar.
Made the selected main menu option background an linear gradient
corporate orange.  Also made the selected submenu a linear gradient
corporate blue.

The rest of the menu is the standard background linear gradient grey.

Changed the font sizes for the main, submenu and items in descending
size and also indented the submenu text to make the structure
more obvious.

Change-Id: Ic7502735572a66f503f8eeb07ce88ac0ea618836
Closes-Bug: DBAAS-498