Add support for root enable and show
Added an entry in the instance details panel to show whether root has ever
been enabled for the instance.

Added a manage root access screen to allow the user to enable the root
user or reset the root password.

The manage root access screen is accessed from a row menu option in the
database instances table.

Change-Id: I93ef2497eacfd0fad5d33bfa65941b9e2a9b95f7
Implements: blueprint trove-enable-root-support