Fix brand-danger error in Tesora theme
A recent upstream fix pulled out a number of variables so that the
theme can customize each one.

The Tesora theme needs to include the default theme as the base.

Also removed the definition of the "regular" variables that were
defined to use the defaults which are darker and more readable.

Change-Id: I18f0724293ee8cfc49a17b54bde894c05755a813
Closes-Bug: DBAAS-697