Change launch instance volume type to pulldown
In the launch instance dialog changed the volume type field to a
pulldown.  It will be populated by a call to cinder volume_type_list.
The default value will be prepended to the list as "no_type".  This
default is used in the cinder panels.

Updated tests to account for new cinder calls and add the default
value to the post commands in the tests.

Change-Id: I461dfe2a710a2b6f712183f29f4cc5fbe866610e