Updated log ui to reflect new implementation
Updated the log table columns to account for new/changed columns.
Changed the "Stop Collection" action name to "Clear Container".
Removed the allow methods for the "Publish" and "Clear Container"
actions so that they are always available
Added an allow method for "View Log" action to only be visible if the
published count is greater than zero.
Changed the default on view log to not automatically publish.
Changed to use the proper api call for the swift client.

Change-Id: I0e00cafef8b1cbef61157ad5670b1f02c3ae5069