Cisco plugin includes tesora-pythontroveclient
If the source branch is dev/EE-1.6 or master the Cisco plugin
now includes the tesora-python-troveclient code.

This is to ensure no requirements issues by installing the
packaged tesora-python-troveclient.

Altered the from troveclient and import troveclient to be
prefixed with tesora_horizon.

Also changed oslo_utils call to the appropriate Juno version.

Change-Id: I242588d338596ef00232c026aebeb2096300ea30