Get all instances to populate master pulldown
The call to gather all instances to populate the master pulldown in
the launch database instance wizard will never exceed the default
pagination size because the call is only made once and never checks
if there is more data.

Looped the call to obtain instances until there is no more data to
retrieve.

Change-Id: If044040901716784277ca52108e83c1a5296fd9c
Closes-bug: DBAAS-1085