Get all instances to populate master pulldown
The call to gather all instances to populate the master pulldown in
the launch database instance wizard will never exceed the default
pagination size because the call is only made once and never checks
if there is more data.

Looped the call to obtain instances until there is no more data to
retrieve.

Change-Id: If330f1975553a04c5d6e74a37c8e98f58f704ce9
Closes-bug: DBAAS-1085