Add support for upgrade instance
Added a row menu action for upgrade instance only if there are
multiple datastore versions for the instance datastore.

The action will bring up a dialog showing the current version
and a dropdown listing all the datastore versions (with the
current version filterd out of the list).

Added a new trove api for instance_upgrade.

Change-Id: I44982d5d6c7550f45fadaff91354d0d7ed608bec
Closes-Bug: DBAAS-1203