Use MySQL qualified names for user actions
Added code in db_capability to provide a fully qualified MySQL name
if the datastore is MySQL (or a variant like MariaDB).

User actions will call the db_capability code to get the name as it
exists now or the new fully qualified MySQL name as determined by
the datastore.

Change-Id: I1959f7b50c0996b6363f139ce0f43d5e28677523
Closes-Bug: DBAAS-1439