Account for uppercase characters in fieldname
The datastore flavor fields are generated per datastore and displayed
when the appropriate datastore is selected in the launch instance
dialog.

The names of these flavor fields are concatenated from the datastore
name and datastore version name.  If any of these names are
uppercased then the field is not shown.

The solution is the hex encode the string (which is lowercased and
numbers) as the field name.  The encoded string can be decoded to
obtain the original datastore and version names.

Added methods to do the encoding and decoding to the launch wizard.

Updated the tests to reflect the new field name.

Change-Id: I6bb8d0d734440bce4c48734a074a8124c3b94a9a
Closes-Bug: DBAAS-1535