Fix the active attribute check
The datastore version object only contains the active attribute
when the user is an admin user.  The datastore version object for
non-admin users does not have the active attribute.  The absence
of the active attribute should be considered active.

Fixed the active attribute check to account for the case when the
attribute is missing.

Also modified the test data to create a datastore version without
the active attribute to test this scenario.

Change-Id: Ibb18b136612e6b01ab42b3c2fe95d6151a7c25c1
Closes-Bug: DBAAS-1654